Added more functionality.
Added Tiles and TIlebar classes but not yet implemented fully, still need to add the updates, renders and controls for the use of the tiles and tilebars, also need to add the textures for the tiles we will be using to create levels. Moves some classes around.  Added new textures for doors and exits doors