Some updates.
Added checkCollision in helper. Font resize for options class. made it bigger. Referenced helper class in options class, added function in particles class to spawn bullets specifically for the player. Added way to save and load Pirates, and also set their base stats when being loaded. Added  "use key' for the player when interacting with doors. Cleaning code when instantiating the tilebar in the world. Added comment for editor controls in PlayGame class.